Build system: better check for XML backend selection
Selecting the XML backend is done through the use of two options:
--with-roxml and --with-libxml2. If both are selected, we choose
libroxml ; if none is selected, we force the selection of libxml2.
The patch corrects a bug in the case where both backend are selected
and one (or both) backends are pointing to a directory (i.e.
use_roxml and/or use_libxml2 are not 'yes' but some other value).
Since we tested against 'yes', we failed to recognize a valid
user choice and we forced the selection of libxml2.