XML backend: support for expat XML parser
This patch adds support for the expat XML parser. Expat is
small enough to be used in the embedded world, and mature
enough to be used in a production environment. It uses a
SAX-like parsing mechanism that needs to be transformed into
a DOM-like parser to be used with klish.