XML backend: fix assertion tests
Both name and help can be NULL if the underlying XML
backend says so (i.e. it fails to find the corresponding
attribute). We shall not test the first character of the
attribute but its existence.

This patch reverts a forgotten change that comes from
the first implementation of the XML backends.