contrib: update openwrt package definition
Update the package definition to use a more recent
version of klish (version 2.0.4 instead of the
antiquated version 1.5.4).

The newer versions are using an XML backend to
replace the integrated C++-based TinyXML. In our
case we select libroxml as it's a base package of
the latest versions of openwrt.